Added more translations. Created new vectors for spanish logo translation. Created a new map maker based upon a vector of a pizza slice. Added script to help automate the JS build process. Removed unused fonts and other files. Added a pushable sidebar so that it menu navigation is consistent across different device sizes. Added admin panel for submitting character pics.

+++ database/migrations/2025_06_20_000010_create_answers_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->decimal('lng', 16, 8)->nullable()->default(null);
             $table->integer('hints_used')->default(0);
             $table->integer('correct')->nullable()->default(null);
-            $table->integer('margin_of_error')->default(0);
+            $table->float('margin_of_error', 11, 8)->default(0);
             $table->timestamps();
 
             $table->foreign('quiz_id')->references('id')->on('quizzes');
